+++ src/data/DefenseData.ts
@@ -10,6 +10,8 @@ export interface DefenseDataResponse {
     baseDefHealth: number;
     baseAtkRate: number;
     maxAtkRate: number;
+    baseRange: number;
+    maxRange: number;
     baseAtkRange: number;
     maxAtkRange: number;
     rangeScalar: number;
@@ -26,6 +28,7 @@ export interface DefenseDataResponse {
     ascDefPwr: number;
     ascDefHp: number;
     ascGambit: number;
+    defenseUnits: number;
     isUnique: boolean;
 }
 
@@ -37,12 +40,16 @@ export default class DefenseData extends HasAscensionPoints implements DefenseRo
     baseDefenseHealth: number;
     baseAttackRate: number;
     maxAttackRate: number;
+    baseRange: number;
+    maxRange: number;
     baseAttackRange: number;
     maxAttackRange: number;
     rangeScalar: number;
     attackScalar: number[];
     hpScalar: number[];
     ascensionPoints: AscensionPointInterface[] = [];
+    hero: string;
+    defenseUnits: number;
     isUnique: boolean = false;
 
     constructor(data: DefenseDataResponse) {
@@ -68,9 +75,13 @@ export default class DefenseData extends HasAscensionPoints implements DefenseRo
         this.baseDefenseHealth = data.baseDefHealth;
         this.baseAttackRate = data.baseAtkRate;
         this.maxAttackRate = data.maxAtkRate;
+        this.baseRange = data.baseRange;
+        this.maxRange = data.maxRange;
         this.baseAttackRange = data.baseAtkRange;
         this.maxAttackRange = data.maxAtkRange;
         this.rangeScalar = data.rangeScalar;
+        this.hero = data.hero;
+        this.defenseUnits = data.defenseUnits;
         this.isUnique = data.isUnique;
         this.attackScalar = [
             data.t1AtkScalar,