+++ src/interaces.ts
@@ -52,6 +52,7 @@ export interface AscensionPointInterface {
     defensePower?: number;
     defenseHealth?: number;
     defenseRate?: number;
+    defenseRange?: number;
     setUpgradeLevel: (level: number) => AscensionPointInterface;
 }
 
@@ -64,12 +65,16 @@ export interface DefenseRootInterface {
     baseDefenseHealth: number;
     baseAttackRate: number;
     maxAttackRate: number;
+    baseRange: number;
+    maxRange: number;
     baseAttackRange: number;
     maxAttackRange: number;
     rangeScalar: number;
     ascensionPoints: AscensionPointInterface[];
     attackScalar: number[];
     hpScalar: number[];
+    hero: string;
+    defenseUnits: number;
     isUnique: boolean;
 }