+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [v1.2.0](https://github.com/kabamgamer/dd2-strategy-hub/tree/v1.2.0) (2023-08-31)
+
+* Added DU calculator to defense setups
+* Added defense range calculator
+* Added defense health calculator
+* Fixed a bug in the UI causing collapse to snap instead of transition
+
+[Full Changelog](https://github.com/kabamgamer/dd2-strategy-hub/compare/v1.1.8...v1.2.0)
+
+
+
 ## [v1.1.8](https://github.com/kabamgamer/dd2-strategy-hub/tree/v1.1.8) (2023-08-29)
 
 * Hide shards/mods if they are incompatible with the defense